BETA

+++ RATAC.C
@@ -2,12 +2,12 @@
 
 int main(void)
 {
-	double b,h,area;
+	long double b,h,area;
 	printf("Right Angle Triangle Area Calculator\n");
 	printf("--------------------------------------------\n");
 	printf("Please Input Base and Height: ");
-	scanf("%f%f",&b,&h);
+	scanf("%Lf%Lf",&b,&h);
 	area = 0.5 * b * h;
-	printf("The Area is %f\n",area);
+	printf("The Area is %Lf\n",area);
 	return 0;
 }
