+++ prime.c
@@ -2,9 +2,9 @@
 #include <conio.h>
 
 void main(void) {
-    int i, num, temp;
+    unsigned long long int i, num, temp;
     printf("Enter : ");
-    scanf("%d",&num);
+    scanf("%llu",&num);
     for (i = 2; i < num; i++) {
         temp = num % i;
     }
